changed table pricing JS and credit card CSS
changed the javascript in the pricing table: I had an arrow function for each radio button and the functions were doing practically the same: changing the prices. I realized the problem with that: I was creating and calling two different functions to do almost the same. So, I decided to create only one function with an if-else statement to check each radio button.

Then,  I added a click event to each card to make them redirect to the credit card page. I realized that I was making the same mistake that I did earlier and corrected it in the same way, creating only one function. But another problem  came to my view: when I go to the credit card page and then back to the pricing table, the prices displaying were not correct. Therefore I added a last event, this time to the document object itself, so when the DOM content is loaded it will check which radio button is checked and display the prices accordingly.

Changed CSS on the credit card:

I had a transform property with a translateY function, and this was working great for the positioning of the page, but it didt work out when I added the animations. So I erased that property and instead I made the position of the credit card relative and this make the card maintain the same position as before without interfering with the animations